# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FinGo API — Obras em Fluxo",
    "version": "2.0.0",
    "description": "API corporativa para gestão financeira, orçamentária, obras de engenharia, NFe/impostos e inteligência operacional.",
    "contact": {
      "name": "Suporte FinGo",
      "url": "https://fingo.api.br",
      "email": "suporte@fingo.api.br"
    },
    "license": {
      "name": "Proprietary"
    }
  },
...
  "servers": [
    {
      "url": "https://fingo.api.br/api",
      "description": "FinGo Edge API Gateway"
    }
  ],
  "paths": {
...
    "/db": {
      "get": {
        "summary": "Consulta de entidades do tenant",
        "description": "Recupera dados de obras, transações, fornecedores, clientes e categorias.",
        "parameters": [
          {
            "name": "entity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Nome da entidade (ex: transacoes, obras, fornecedores)."
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de registros retornada."
          },
          "401": {
            "description": "Sessão não informada ou inválida."
          }
        }
      },
      "post": {
        "summary": "Mutações e sincronização de dados",
        "description": "Criação, edição e exclusão de registros financeiros e de engenharia.",
        "responses": {
          "200": {
            "description": "Registro salvo com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          }
        }
      }
    },
...
    "/plano": {
      "get": {
        "summary": "Consultar status e limites do plano",
        "responses": {
          "200": {
            "description": "Informações do plano retornadas."
          }
        }
      },
      "post": {
        "summary": "Checkout e alteração de plano",
        "x-payment-info": {
          "intent": "charge",
          "method": "card",
          "amount": "119.90",
          "currency": "BRL",
          "description": "Assinatura mensal do Plano Básico FinGo"
        },
        "responses": {
          "200": {
            "description": "Transação de assinatura processada."
          },
          "402": {
            "description": "Payment Required via protocolo x402 ou checkout tradicional."
          }
        }
      }
    },
...
  }
}